requirements.txt
paho-mqtt
signalrcore
turfpy==0.0.7